moving getReparentTargetForFlexElement to helper file
